requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
Werkzeug==2.0.3
